fix: 修改HomePageShowSoftwareDTO

+++ midExam/SoftwarePlatform/src/main/java/com/qg24/softwareplatform/po/dto/HomePageShowSoftwareDTO.java
@@ -13,7 +13,11 @@ public class HomePageShowSoftwareDTO {
     /**
      * 页码
      */
-    private long page;
+    private int page;
+
+    /**
+     *
+     */
     /**
      * 软件名
      */
